requirements.txt

torch>=2.1.0
torchvision
torchaudio


transformers>=4.38.0
datasets>=2.16.0
accelerate>=0.27.0
bitsandbytes>=0.42.0  # 用于 4-bit 量化加载


peft>=0.8.2  # LoRA
trl>=0.7.10  # SFTTrainer


pyyaml
openai>=1.12.0
tqdm
scipy
sentencepiece
protobuf

tensorboard
wandb